+++ common/constants.ts
@@ -1,0 +1,3 @@
+export const TOTAL_PRODUCTS = 'total';
+export const BLOB_CONTAINER_NAME = 'uploaded';
+export const PARSE_CONTAINER_NAME = 'parsed';