+++ terraform/main.tf
@@ -1,3 +1,5 @@
+data "azurerm_client_config" "current" {}
+
 provider "azurerm" {
   features {}
 }
@@ -159,6 +161,7 @@ resource "azurerm_api_management_api_policy" "api_policy" {
             </allowed-origins>
             <allowed-methods>
                 <method>GET</method>
+                <method>POST</method>
             </allowed-methods>
             <allowed-headers>
                 <header>*</header>
@@ -212,10 +215,206 @@ resource "azurerm_api_management_api_operation" "get_product_by_id" {
   }
 }
 
+resource "azurerm_api_management_api_operation" "post_product" {
+  api_management_name = azurerm_api_management.core_apim.name
+  api_name            = azurerm_api_management_api.products_api.name
+  display_name        = "Post Product"
+  method              = "POST"
+  operation_id        = "post-product"
+  resource_group_name = azurerm_resource_group.apim.name
+  url_template        = "/products"
+}
+
+resource "azurerm_api_management_api_operation" "health" {
+  api_management_name = azurerm_api_management.core_apim.name
+  api_name            = azurerm_api_management_api.products_api.name
+  display_name        = "Health"
+  method              = "GET"
+  operation_id        = "health"
+  resource_group_name = azurerm_resource_group.apim.name
+  url_template        = "/health"
+}
+
 resource "azurerm_app_configuration" "products_config" {
   location            = "northeurope"
   name                = "appconfig-products-service-sand-ne-123"
   resource_group_name = azurerm_resource_group.product_service_rg.name
 
   sku = "free"
+}
+
+resource "azurerm_cosmosdb_account" "cosmosdb_account" {
+  name                = "cosmosdb-account-sand-ne-123"
+  location            = "northeurope"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  offer_type          = "Standard"
+
+  consistency_policy {
+  consistency_level       = "BoundedStaleness"
+  max_interval_in_seconds = 300
+  max_staleness_prefix    = 100000
+  }
+
+  geo_location {
+    location          = "northeurope"
+    failover_priority = 0
+  }
+}
+
+resource "azurerm_cosmosdb_sql_database" "cosmos_database" {
+  name                = "cosmosdb-database-sand-ne-123"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  account_name        = azurerm_cosmosdb_account.cosmosdb_account.name
+}
+
+resource "azurerm_cosmosdb_sql_container" "product" {
+  name                = "Product"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  account_name        = azurerm_cosmosdb_account.cosmosdb_account.name
+  database_name       = azurerm_cosmosdb_sql_database.cosmos_database.name
+  partition_key_path    = "/id"
+
+  indexing_policy {
+    indexing_mode = "consistent"
+  }
+}
+
+resource "azurerm_cosmosdb_sql_container" "stock" {
+  name                = "Stock"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  account_name        = azurerm_cosmosdb_account.cosmosdb_account.name
+  database_name       = azurerm_cosmosdb_sql_database.cosmos_database.name
+  partition_key_path    = "/id"
+
+  indexing_policy {
+    indexing_mode = "consistent"
+  }
+}
+
+resource "azurerm_cosmosdb_sql_role_definition" "read_write_role" {
+  name                = "PasswordlessReadWrite"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  account_name        = azurerm_cosmosdb_account.cosmosdb_account.name
+  type                = "CustomRole"
+  assignable_scopes   = [azurerm_cosmosdb_account.cosmosdb_account.id]
+
+  permissions {
+    data_actions = ["Microsoft.DocumentDB/databaseAccounts/readMetadata",
+            "Microsoft.DocumentDB/databaseAccounts/sqlDatabases/containers/items/*",
+            "Microsoft.DocumentDB/databaseAccounts/sqlDatabases/containers/*"]
+  }
+}
+
+resource "azurerm_cosmosdb_sql_role_assignment" "read_write_role_assignment" {
+  name                = "736180af-7fbc-4c7f-9004-22735173c1c3"
+  resource_group_name = azurerm_resource_group.product_service_rg.name
+  account_name        = azurerm_cosmosdb_account.cosmosdb_account.name
+  role_definition_id  = azurerm_cosmosdb_sql_role_definition.read_write_role.id
+  principal_id        = "f321e0f2-ee2f-4919-8b65-7ab3f41c1fa7"
+  scope               = azurerm_cosmosdb_account.cosmosdb_account.id
+}
+
+resource "azurerm_resource_group" "import_service_rg" {
+  location = "northeurope"
+  name     = "rg-product-service"
+}
+
+resource "azurerm_storage_account" "sa" {
+  name                             = "stgsangproductsfane123"
+  resource_group_name              = azurerm_resource_group.import_service_rg.name
+  location                         = azurerm_resource_group.import_service_rg.location
+  account_tier                     = "Standard"
+  account_replication_type         = "LRS"
+  access_tier                      = "Cool"
+  enable_https_traffic_only        = true
+  allow_nested_items_to_be_public  = true
+  shared_access_key_enabled        = true
+}
+
+resource "azurerm_storage_container" "sa_container" {
+  name                  = "upload-container"
+  storage_account_name  = azurerm_storage_account.sa.name
+  container_access_type = "private"
+}
+
+resource "azurerm_storage_blob" "sa_blob" {
+  name                   = "import-content.zip"
+  storage_account_name   = azurerm_storage_account.sa.name
+  storage_container_name = azurerm_storage_container.sa_container.name
+  type                   = "Block"
+}
+
+resource "azurerm_storage_share" "import_service_fa" {
+  name  = "fa-import-service-share"
+  quota = 2
+
+  storage_account_name = azurerm_storage_account.sa.name
+}
+
+resource "azurerm_service_plan" "import_service_plan" {
+  name     = "asp-import-service-sand-ne-001"
+  location = "northeurope"
+
+  os_type  = "Windows"
+  sku_name = "Y1"
+
+  resource_group_name = azurerm_resource_group.import_service_rg.name
+}
+
+resource "azurerm_application_insights" "import_service_fa" {
+  name             = "appins-fa-import-service-sand-ne-001"
+  application_type = "web"
+  location         = "northeurope"
+
+
+  resource_group_name = azurerm_resource_group.import_service_rg.name
+}
+
+resource "azurerm_windows_function_app" "import_service" {
+  name     = "fa-import-service-ne-123"
+  location = "northeurope"
+
+  service_plan_id     = azurerm_service_plan.import_service_plan.id
+  resource_group_name = azurerm_resource_group.import_service_rg.name
+
+  storage_account_name       = azurerm_storage_account.sa.name
+  storage_account_access_key = azurerm_storage_account.sa.primary_access_key
+
+  functions_extension_version = "~4"
+  builtin_logging_enabled     = false
+
+  site_config {
+    always_on = false
+
+    application_insights_key               = azurerm_application_insights.import_service_fa.instrumentation_key
+    application_insights_connection_string = azurerm_application_insights.import_service_fa.connection_string
+
+    # For production systems set this to false
+    use_32_bit_worker = true
+
+    # Enable function invocations from Azure Portal.
+    cors {
+      allowed_origins = ["https://portal.azure.com", "*"]
+    }
+
+    application_stack {
+      node_version = "~16"
+    }
+  }
+
+  app_settings = {
+    WEBSITE_CONTENTAZUREFILECONNECTIONSTRING = azurerm_storage_account.sa.primary_connection_string
+    WEBSITE_CONTENTSHARE                     = azurerm_storage_share.import_service_fa.name
+  }
+
+  # The app settings changes cause downtime on the Function App. e.g. with Azure Function App Slots
+  # Therefore it is better to ignore those changes and manage app settings separately off the Terraform.
+  lifecycle {
+    ignore_changes = [
+      app_settings,
+      tags["hidden-link: /app-insights-instrumentation-key"],
+      tags["hidden-link: /app-insights-resource-id"],
+      tags["hidden-link: /app-insights-conn-string"]
+    ]
+  }
 }